New article service

+++ src/EFPersistenceModel/ArticleRepository.cs
@@ -37,7 +37,7 @@
             return articles.Select(x => x.ToDomain());
         }
 
-        public Task<DomainArticle> FineAsync(params object[] identity)
+        public Task<DomainArticle> FindAsync(params object[] identity)
         {
             if (identity == null)
                 throw new ArgumentNullException("identity");
@@ -71,8 +71,12 @@
             return Task.FromResult<object>(null);
         }
 
-        public Task DeleteAsync(int id)
+        public Task DeleteAsync(params object[] identity)
         {
+            if (identity == null)
+                throw new ArgumentNullException("identity");
+
+            var id = (int)identity[0];
             var article = this.context.Articles.Find(id);
             if (article != null)
                 this.context.Articles.Remove(article);
@@ -80,17 +84,9 @@
             return Task.FromResult<object>(null);
         }
 
-        public Task DeleteAsync(params object[] identity)
-        {
-            if (identity == null)
-                throw new ArgumentNullException("identity");
-
-            throw new NotImplementedException();
-        }
-
         private async Task<DomainArticle> InsertAsyncImpl(DomainArticle item)
         {
-            if ((await this.FineAsync(new object[] { item.Id })) != null)
+            if ((await this.FindAsync(new object[] { item.Id })) != null)
                 return item;
 
             var persistenceArticle = item.ToPersistence(await this.GetUserId(item));
